+++ next/binary/include/binary/Converter.hpp
@@ -4,7 +4,6 @@
 #define BINARY_CONVERTER_HPP
 
 #include <cstddef>
-#include <cstdint>
 #include <span>
 
 #include "Allocator.hpp"
@@ -18,13 +17,13 @@ public:
     Converter()
         : IConverter(typeid(T)) {}
 
-    Converter(int32_t length)
+    Converter(size_t length)
         : IConverter(typeid(T), length) {}
 
     virtual void Encode(Allocator& allocator, const T& item) = 0;
 
     virtual void EncodeAuto(Allocator& allocator, const T& item) {
-        int32_t length = Length();
+        size_t length = Length();
         if (length != 0) {
             Encode(allocator, item);
         } else {
@@ -33,7 +32,7 @@ public:
     }
 
     virtual void EncodeWithLengthPrefix(Allocator& allocator, const T& item) {
-        int32_t anchor = allocator.Anchor();
+        size_t anchor = allocator.Anchor();
         Encode(allocator, item);
         allocator.FinishAnchor(anchor);
     }
@@ -41,10 +40,10 @@ public:
     virtual T Decode(const std::span<const std::byte>& span) = 0;
 
     virtual T DecodeAuto(std::span<const std::byte>& span) {
-        int32_t length = Length();
+        size_t length = Length();
         if (length != 0) {
             T result = Decode(span);
-            span = span.subspan(static_cast<size_t>(length));
+            span = span.subspan(length);
             return result;
         } else {
             return DecodeWithLengthPrefix(span);
